pyproject.toml
[tool.poetry]
name = "pimviback"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
prisma = "^0.8.2"
pytest = "^7.3.1"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.0"
faker = "^18.9.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.pytest.ini_options]
pythonpath = [
    "."
]
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
